requirements.txt
kfp==2.0.1
google-cloud-aiplatform==1.33.0
google-cloud-pipeline-components==2.3.1
gcsfs==2023.6.0

